update comments and some reorganization of files (IDReferenced -> Database package)
