+++ src/main/java/se/uog/App.java
@@ -9,6 +9,7 @@ public class App {
 
     public static void main(String[] args) {
 
+        // Logger, for debugging if needed.
         if (args.length > 0 && args[0].equals(LOGGER_FLAG)) {
             System.out.println("Logger enabled");
             Logger.getInstance().enable();
