+++ src/main/java/se/uog/database/IDReferenced.java
@@ -1,7 +1,11 @@
-package se.uog;
+package se.uog.database;
 
 import java.util.UUID;
 
+/**
+ * This abstract class gives classes that extend from it an ID which is employed by the model
+ * serializer / deserializer.
+ */
 public abstract class IDReferenced {
 
     private String id;
